Expand saag-aloo.html content and fix typo
Added a description to the saag-aloo.html recipe
page and started added in unordered list of
ingredients.

Spotted a glaring typo, where I had used sag
instead of saag in all filenames, titles and
text. Addressed this by updating all occurrences
of the word and testing to ensure website still
functions as expected.