Improve index.html appearance and add introduction
Improvements made to index.html main page to make
it more readable and user friendly.
Image added to images directory and used on the
index.html page for decorative purposes.
Added a short introduction to the website under
the image.
Created an unordered list to house the links to
the recipe pages on the site, making them easier
to tell apart.